split vkengine in separate file
